+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "rocket-code" extension will be documented in this fi
 ## 0.0.2 [Unreleased]
 - Changed activation event from "*" to login()
 - Added icon for extension
+- List joined #channels
+- Select current #channel
 
 ## 0.0.1 [Unreleased]
 - Implement basic Rocket.Chat API access